Fuse annotations into the type-erased function trampoline
The type-erased  trampoline generated by ``.def()`` function bindings
previously received a per-argument flags array ``args_flags`` and a runtime return
value policy (``rvp``) from the dispatcher.

```cpp
PyObject *(void *ptr, PyObject **args, uint8_t *args_flags, rv_policy rvp, cleanup_list *)
```

Both are now baked into the generated code at compile time, and the
signature reduces to

```cpp
PyObject *(void *ptr, PyObject **args, uint32_t flags, cleanup_list *)
```

with a per-call (and not per-argument) flag. Flags are now consistently
32-bit since that comes at no cost, though nothing in the implementation
uses those extra bits for now.

The argument descriptors collapse into ``arg_t<Flags, Locked>`` and
``arg_v_t<Flags, Locked>`` (with a default value) which the trampoline
queries at compile time. It is no longer legal to pass booleans to
argument binding annotations. For example,

```cpp
m.def("f", &f, "arg"_a.none(true))
```

must change to

```cpp
m.def("f", &f, "arg"_a.none())
```

or

```cpp
m.def("f", &f, "arg"_a = nb::none())
```

The commit also changes :cpp:class:`nb::none <none>` to be a type, whose
constructor generates the ``None`` singleton. This enables detection of
the above pattern at compile time.

Finally, the flags word of the type caster interface widens from
``uint8_t`` to ``uint32_t``: ``from_python()`` implementations now
receive a ``uint32_t``. A later commit inserts an additional
``nb_internals *`` parameter into the same interface.

+++ include/nanobind/eigen/dense.h
@@ -134,11 +134,11 @@ struct type_caster<T, enable_if_t<is_eigen_plain_v<T> &&
 
     NB_TYPE_CASTER(T, NDArrayCaster::Name)
 
-    bool from_python(handle src, uint8_t flags, cleanup_list *cleanup) noexcept {
+    bool from_python(handle src, uint32_t flags, cleanup_list *cleanup) noexcept {
         // We're in any case making a copy, so non-writable inputs area also okay
         using NDArrayConst = array_for_eigen_t<T, const typename T::Scalar>;
         make_caster<NDArrayConst> caster;
-        if (!caster.from_python(src, flags & ~(uint8_t)cast_flags::accepts_none, cleanup))
+        if (!caster.from_python(src, flags & ~cast_flags::accepts_none, cleanup))
             return false;
 
         const NDArrayConst &array = caster.value;
@@ -220,7 +220,7 @@ struct type_caster<T, enable_if_t<is_eigen_xpr_v<T> &&
     template <typename T_> static constexpr bool can_cast() { return true; }
 
     /// Generating an expression template from a Python object is, of course, not possible
-    bool from_python(handle src, uint8_t flags, cleanup_list *cleanup) noexcept = delete;
+    bool from_python(handle src, uint32_t flags, cleanup_list *cleanup) noexcept = delete;
 
     template <typename T2>
     static handle from_cpp(T2 &&v, rv_policy policy, cleanup_list *cleanup) noexcept {
@@ -252,13 +252,13 @@ struct type_caster<Eigen::Map<T, Options, StrideType>,
 
     NDArrayCaster caster;
 
-    bool from_python(handle src, uint8_t flags, cleanup_list *cleanup) noexcept {
+    bool from_python(handle src, uint32_t flags, cleanup_list *cleanup) noexcept {
         // Disable implicit conversions
-        return from_python_(src, flags & ~(uint8_t)cast_flags::convert, cleanup);
+        return from_python_(src, flags & ~cast_flags::convert, cleanup);
     }
 
-    bool from_python_(handle src, uint8_t flags, cleanup_list* cleanup) noexcept {
-        if (!caster.from_python(src, flags & ~(uint8_t)cast_flags::accepts_none, cleanup))
+    bool from_python_(handle src, uint32_t flags, cleanup_list* cleanup) noexcept {
+        if (!caster.from_python(src, flags & ~cast_flags::accepts_none, cleanup))
             return false;
 
         // Check for memory layout compatibility of non-contiguous 'Map' types
@@ -425,7 +425,7 @@ struct type_caster<Eigen::Ref<T, Options, StrideType>,
     struct Empty { };
     std::conditional_t<MaybeConvert, DMapCaster, Empty> dcaster;
 
-    bool from_python(handle src, uint8_t flags, cleanup_list *cleanup) noexcept {
+    bool from_python(handle src, uint32_t flags, cleanup_list *cleanup) noexcept {
         // Try a direct cast without implicit conversion first
         if (caster.from_python(src, flags, cleanup))
             return true;
@@ -443,9 +443,9 @@ struct type_caster<Eigen::Ref<T, Options, StrideType>,
                If neither of these is possible, we disable implicit
                conversions. */
 
-            if ((flags & (uint8_t) cast_flags::manual) &&
+            if ((flags & cast_flags::manual) &&
                 !DMapConstructorOwnsData)
-                flags &= ~(uint8_t) cast_flags::convert;
+                flags &= ~cast_flags::convert;
 
             if (dcaster.from_python_(src, flags, cleanup))
                 return true;
