+++ include/nanobind/stl/complex.h
@@ -16,12 +16,12 @@
 NAMESPACE_BEGIN(NB_NAMESPACE)
 NAMESPACE_BEGIN(detail)
 
-NB_CORE bool load_cmplx(PyObject*, uint8_t flags, std::complex<double> *out) noexcept;
+NB_CORE bool load_cmplx(PyObject*, uint32_t flags, std::complex<double> *out) noexcept;
 
 template <typename T> struct type_caster<std::complex<T>> {
     NB_TYPE_CASTER(std::complex<T>, const_name("complex"))
 
-    bool from_python(handle src, uint8_t flags, cleanup_list*) noexcept {
+    bool from_python(handle src, uint32_t flags, cleanup_list*) noexcept {
         std::complex<double> cmplx;
         if (!load_cmplx(src.ptr(), flags, &cmplx))
             return false;
@@ -31,7 +31,7 @@ template <typename T> struct type_caster<std::complex<T>> {
         } else {
             T re = (T) cmplx.real();
             T im = (T) cmplx.imag();
-            if ((flags & (uint8_t) cast_flags::convert)
+            if ((flags & cast_flags::convert)
                     ||  (((double) re == cmplx.real()
                              || (re != re && cmplx.real() != cmplx.real()))
                       && ((double) im == cmplx.imag()
