+++ include/nanobind/stl/detail/nb_array.h
@@ -12,7 +12,7 @@ template <typename Array, typename Entry, size_t Size> struct array_caster {
 
     using Caster = make_caster<Entry>;
 
-    bool from_python(handle src, uint8_t flags, cleanup_list *cleanup) noexcept {
+    bool from_python(handle src, uint32_t flags, cleanup_list *cleanup) noexcept {
         PyObject *temp;
 
         /* Will initialize 'temp' (NULL in the case of a failure.) */
