+++ include/nanobind/stl/pair.h
@@ -33,7 +33,7 @@ template <typename T1, typename T2> struct type_caster<std::pair<T1, T2>> {
         const_name("tuple[") + concat(Caster1::Name, Caster2::Name) + const_name("]");
 
     /// Python -> C++ caster, populates `caster1` and `caster2` upon success
-    bool from_python(handle src, uint8_t flags,
+    bool from_python(handle src, uint32_t flags,
                      cleanup_list *cleanup) noexcept {
         PyObject *temp; // always initialized by the following line
         PyObject **o = seq_get_with_size(src.ptr(), 2, &temp);
