+++ include/nanobind/stl/string.h
@@ -18,7 +18,7 @@ NAMESPACE_BEGIN(detail)
 template <> struct type_caster<std::string> {
     NB_TYPE_CASTER(std::string, const_name("str"))
 
-    bool from_python(handle src, uint8_t, cleanup_list *) noexcept {
+    bool from_python(handle src, uint32_t, cleanup_list *) noexcept {
         if (!PyUnicode_Check(src.ptr()))
             return false;
         Py_ssize_t size;
